JsonDocumentGet god class divided into logical units
